New WIP port: Echo R1
The Echo R1 is a new open-hardware music player design, based
on the STM32H743. Schematics and hardware documentation for it
can be found here:

- https://github.com/amachronic/echoplayer

This is an incomplete port. The bootloader can be loaded using
OpenOCD and it can draw to the LCD using SPI. SDRAM is working
but hasn't been extensively tested.

Change-Id: Ifd2bee15c49868fbc989683d3ca14dce48bf3e18